requirements.txt
fastapi
uvicorn
jinja2
pydantic
python-multipart
mysql-connector-python
sqlalchemy
python-dotenv